src/main/java/com/fbm/services/UserService.java
package com.fbm.services;


import com.fbm.domain.User;
import com.fbm.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created by Mocart on 09-Jul-17.
 */

@Service
public class UserService {
    @Autowired
    private UserRepository userRepository;

    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    public User getById(long userId) {
        return userRepository.findOne(userId);
    }

    public User getByUserName(String userName) {
        return userRepository.findByUserName(userName);
    }

    public User addUser(User user) {
        return userRepository.save(user);
    }

    public User updateUser(User user) {
        return userRepository.save(user);
    }

    public void deleteById(long id) {
        userRepository.delete(id);
    }

    public void deleteUser(User user) {
        userRepository.delete(user);
    }
}
